Move PreferredResolver
